+++ src/meltano/core/block/plugin_command.py
@@ -111,6 +111,10 @@ class InvokerCommand(InvokerBase, PluginCommandBlock):
         """
         return self._command_args
 
+    async def _start(self):
+        invoke_args = (self.command_args,) if self.command_args else ()
+        await self.start(*invoke_args)
+
     async def run(self) -> None:
         """Invoke a command capturing and logging produced output.
 
@@ -119,7 +123,7 @@ class InvokerCommand(InvokerBase, PluginCommandBlock):
         """
         try:  # noqa: WPS501
             async with self.invoker.prepared(self.context.session):
-                await self.start(self.command_args)
+                await self._start()
 
                 self.stdout_link(self._log)
                 self.stderr_link(self._log)